+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Technical
+
+- Fixed: `simulation` event are now captured after create and include the persisted simulation `id`.
+
 ## [0.0.2] - 2026-07-16
 
 ### Public features